requirements.txt
Flask
openai
faiss-cpu
PyMuPDF
numpy
